Made a lot of progress
