+++ Lamp/Lamp.vcxproj.filters
@@ -1281,6 +1281,9 @@
     <ClInclude Include="vendor\stb_image\stb\stb_image.h">
       <Filter>vendor\stb_image\stb</Filter>
     </ClInclude>
+    <ClInclude Include="src\Lamp\Entity\Base\ComponentProperties.h">
+      <Filter>src\Lamp\Entity\Base</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\Lamp\Core\Application.cpp">
