+++ Lamp/src/Lamp/Entity/BaseComponents/TransformComponent.h
@@ -10,9 +10,15 @@ namespace Lamp
 	{
 	public:
 		TransformComponent(const glm::vec3& pos)
-			: m_Scale(glm::vec3(1.f, 1.f, 1.f)), m_Rotation(glm::vec3(0, 0, 0))
+			: IEntityComponent("Transform Component"), m_Scale(glm::vec3(1.f, 1.f, 1.f)), m_Rotation(glm::vec3(0, 0, 0))
 		{
 			m_Position = pos;
+			SetComponentProperties
+			({
+				{ PropertyType::Float3, "Position", static_cast<void*>(&m_Position) },
+				{ PropertyType::Float3, "Rotation", static_cast<void*>(&m_Rotation) },
+				{ PropertyType::Float3, "Scale", static_cast<void*>(&m_Scale) }
+			});
 		}
 
 		virtual void Initialize() override {}
@@ -23,7 +29,6 @@ namespace Lamp
 		inline const glm::vec3& GetPosition() const { return m_Position; }
 		inline const glm::vec3& GetRotation() const { return m_Rotation; }
 		inline const glm::vec3& GetScale() const { return m_Scale; }
-		virtual const EditorValues GetEditorValues() const { return m_TransformValues; }
 
 		//Setting
 		inline void SetPosition(const glm::vec3& pos) { m_Position = pos; }
@@ -34,14 +39,5 @@ namespace Lamp
 		glm::vec3 m_Position;
 		glm::vec3 m_Rotation;
 		glm::vec3 m_Scale;
-
-		struct TransformValues : EditorValues
-		{
-			TransformValues()
-				: EditorValues("Transform Component")
-			{}
-		};
-
-		TransformValues m_TransformValues = TransformValues();
 	};
 }
